+++ x/profiles/handler.go
@@ -64,6 +64,14 @@ func NewHandler(k keeper.Keeper) sdk.Handler {
 			res, err := msgServer.UnlinkChainAccount(sdk.WrapSDKContext(ctx), msg)
 			return sdk.WrapServiceResult(ctx, res, err)
 
+		case *types.MsgLinkApplication:
+			res, err := msgServer.LinkApplication(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
+
+		case *types.MsgUnlinkApplication:
+			res, err := msgServer.UnlinkApplication(sdk.WrapSDKContext(ctx), msg)
+			return sdk.WrapServiceResult(ctx, res, err)
+
 		default:
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest,
 				"unrecognized %s message type: %v", types.ModuleName, msg.Type())