+++ x/profiles/types/query_chain_links.pb.go
@@ -27,6 +27,118 @@ var _ = math.Inf
 // proto package needs to be updated.
 const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
+// QueryUserChainLinkRequest represents the request that should be used in order
+// to retrieve the link associated with the provided user, for the given chain
+// and having the given target address
+type QueryUserChainLinkRequest struct {
+	// User represents the Desmos address of the user to which search the link for
+	User string `protobuf:"bytes,1,opt,name=user,proto3" json:"user,omitempty"`
+	// ChainName contains the name of the chain to which search the link for
+	ChainName string `protobuf:"bytes,2,opt,name=chain_name,json=chainName,proto3" json:"chain_name,omitempty"`
+	// Target must contain the external address to which query the link for
+	Target string `protobuf:"bytes,3,opt,name=target,proto3" json:"target,omitempty"`
+}
+
+func (m *QueryUserChainLinkRequest) Reset()         { *m = QueryUserChainLinkRequest{} }
+func (m *QueryUserChainLinkRequest) String() string { return proto.CompactTextString(m) }
+func (*QueryUserChainLinkRequest) ProtoMessage()    {}
+func (*QueryUserChainLinkRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_c44c4be38a628772, []int{0}
+}
+func (m *QueryUserChainLinkRequest) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QueryUserChainLinkRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QueryUserChainLinkRequest.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QueryUserChainLinkRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QueryUserChainLinkRequest.Merge(m, src)
+}
+func (m *QueryUserChainLinkRequest) XXX_Size() int {
+	return m.Size()
+}
+func (m *QueryUserChainLinkRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_QueryUserChainLinkRequest.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QueryUserChainLinkRequest proto.InternalMessageInfo
+
+func (m *QueryUserChainLinkRequest) GetUser() string {
+	if m != nil {
+		return m.User
+	}
+	return ""
+}
+
+func (m *QueryUserChainLinkRequest) GetChainName() string {
+	if m != nil {
+		return m.ChainName
+	}
+	return ""
+}
+
+func (m *QueryUserChainLinkRequest) GetTarget() string {
+	if m != nil {
+		return m.Target
+	}
+	return ""
+}
+
+// QueryUserChainLinkResponse contains the data that is returned when querying a
+// specific chain link
+type QueryUserChainLinkResponse struct {
+	Link ChainLink `protobuf:"bytes,1,opt,name=link,proto3" json:"link"`
+}
+
+func (m *QueryUserChainLinkResponse) Reset()         { *m = QueryUserChainLinkResponse{} }
+func (m *QueryUserChainLinkResponse) String() string { return proto.CompactTextString(m) }
+func (*QueryUserChainLinkResponse) ProtoMessage()    {}
+func (*QueryUserChainLinkResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_c44c4be38a628772, []int{1}
+}
+func (m *QueryUserChainLinkResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QueryUserChainLinkResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QueryUserChainLinkResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QueryUserChainLinkResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QueryUserChainLinkResponse.Merge(m, src)
+}
+func (m *QueryUserChainLinkResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *QueryUserChainLinkResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_QueryUserChainLinkResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QueryUserChainLinkResponse proto.InternalMessageInfo
+
+func (m *QueryUserChainLinkResponse) GetLink() ChainLink {
+	if m != nil {
+		return m.Link
+	}
+	return ChainLink{}
+}
+
 // QueryUserChainLinksRequest is the request type for the
 // Query/UserChainLinks RPC endpoint
 type QueryUserChainLinksRequest struct {
@@ -39,7 +151,7 @@ func (m *QueryUserChainLinksRequest) Reset()         { *m = QueryUserChainLinksR
 func (m *QueryUserChainLinksRequest) String() string { return proto.CompactTextString(m) }
 func (*QueryUserChainLinksRequest) ProtoMessage()    {}
 func (*QueryUserChainLinksRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_c44c4be38a628772, []int{0}
+	return fileDescriptor_c44c4be38a628772, []int{2}
 }
 func (m *QueryUserChainLinksRequest) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -94,7 +206,7 @@ func (m *QueryUserChainLinksResponse) Reset()         { *m = QueryUserChainLinks
 func (m *QueryUserChainLinksResponse) String() string { return proto.CompactTextString(m) }
 func (*QueryUserChainLinksResponse) ProtoMessage()    {}
 func (*QueryUserChainLinksResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_c44c4be38a628772, []int{1}
+	return fileDescriptor_c44c4be38a628772, []int{3}
 }
 func (m *QueryUserChainLinksResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -138,6 +250,8 @@ func (m *QueryUserChainLinksResponse) GetPagination() *query.PageResponse {
 }
 
 func init() {
+	proto.RegisterType((*QueryUserChainLinkRequest)(nil), "desmos.profiles.v1beta1.QueryUserChainLinkRequest")
+	proto.RegisterType((*QueryUserChainLinkResponse)(nil), "desmos.profiles.v1beta1.QueryUserChainLinkResponse")
 	proto.RegisterType((*QueryUserChainLinksRequest)(nil), "desmos.profiles.v1beta1.QueryUserChainLinksRequest")
 	proto.RegisterType((*QueryUserChainLinksResponse)(nil), "desmos.profiles.v1beta1.QueryUserChainLinksResponse")
 }
@@ -147,30 +261,111 @@ func init() {
 }
 
 var fileDescriptor_c44c4be38a628772 = []byte{
-	// 368 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x84, 0x92, 0xcf, 0x4e, 0xea, 0x40,
-	0x14, 0xc6, 0x3b, 0xf7, 0x72, 0x6f, 0x62, 0xd9, 0x35, 0x26, 0x42, 0x35, 0x95, 0xb0, 0x50, 0x62,
-	0xe2, 0x8c, 0xe0, 0xde, 0x05, 0x26, 0xea, 0xc2, 0x85, 0x36, 0x71, 0xe3, 0x86, 0x4c, 0xe1, 0x50,
-	0x1a, 0xcb, 0x9c, 0xd2, 0x69, 0x0d, 0xbc, 0x85, 0x4f, 0xe1, 0xb3, 0xb0, 0x64, 0xe9, 0xca, 0x18,
-	0x78, 0x11, 0xd3, 0x99, 0xe1, 0x4f, 0x22, 0xc4, 0xdd, 0x9c, 0x9c, 0xf3, 0x7d, 0xdf, 0x6f, 0x66,
-	0x8e, 0xcd, 0x7a, 0x20, 0x87, 0x28, 0x59, 0x92, 0x62, 0x3f, 0x8a, 0x41, 0xb2, 0xd7, 0x66, 0x00,
-	0x19, 0x6f, 0xb2, 0x51, 0x0e, 0xe9, 0xa4, 0xd3, 0x1d, 0xf0, 0x48, 0x74, 0xe2, 0x48, 0xbc, 0x48,
-	0x9a, 0xa4, 0x98, 0xa1, 0x73, 0xa0, 0x05, 0x74, 0x29, 0xa0, 0x46, 0xe0, 0xee, 0x87, 0x18, 0xa2,
-	0x9a, 0x61, 0xc5, 0x49, 0x8f, 0xbb, 0x47, 0x21, 0x62, 0x18, 0x03, 0xe3, 0x49, 0xc4, 0xb8, 0x10,
-	0x98, 0xf1, 0x2c, 0x42, 0x61, 0xcc, 0xdc, 0xaa, 0xe9, 0xaa, 0x2a, 0xc8, 0xfb, 0x8c, 0x8b, 0x89,
-	0x69, 0x5d, 0xec, 0x02, 0x1b, 0x62, 0x0f, 0x62, 0xf9, 0x93, 0xcc, 0xad, 0x76, 0xb1, 0x50, 0x74,
-	0x34, 0x83, 0x2e, 0x4c, 0xeb, 0x4c, 0x57, 0x2c, 0xe0, 0x12, 0xf4, 0xcd, 0x56, 0x76, 0x09, 0x0f,
-	0x23, 0xa1, 0xa0, 0xf4, 0x6c, 0x7d, 0x6c, 0xbb, 0x8f, 0xc5, 0xc4, 0x93, 0x84, 0xf4, 0xba, 0x08,
-	0xb9, 0x2f, 0x32, 0x7c, 0x18, 0xe5, 0x20, 0x33, 0xc7, 0xb1, 0x4b, 0xb9, 0x84, 0xb4, 0x42, 0x6a,
-	0xa4, 0xb1, 0xe7, 0xab, 0xb3, 0x73, 0x63, 0xdb, 0x6b, 0x97, 0xca, 0x9f, 0x1a, 0x69, 0x94, 0x5b,
-	0x27, 0xd4, 0x00, 0x14, 0x91, 0x54, 0x45, 0x2e, 0x5f, 0x8a, 0x3e, 0xf0, 0x10, 0x8c, 0x9f, 0xbf,
-	0xa1, 0xac, 0xbf, 0x13, 0xfb, 0x70, 0x6b, 0xb4, 0x4c, 0x50, 0x48, 0x70, 0xae, 0xec, 0x7f, 0xea,
-	0xbe, 0x15, 0x52, 0xfb, 0xdb, 0x28, 0xb7, 0xea, 0x74, 0xc7, 0x57, 0xd0, 0x95, 0xb6, 0x5d, 0x9a,
-	0x7e, 0x1e, 0x5b, 0xbe, 0x96, 0x39, 0xb7, 0x5b, 0x38, 0x4f, 0x7f, 0xe5, 0xd4, 0xe1, 0x9b, 0xa0,
-	0xed, 0xbb, 0xe9, 0xdc, 0x23, 0xb3, 0xb9, 0x47, 0xbe, 0xe6, 0x1e, 0x79, 0x5b, 0x78, 0xd6, 0x6c,
-	0xe1, 0x59, 0x1f, 0x0b, 0xcf, 0x7a, 0xa6, 0x61, 0x94, 0x0d, 0xf2, 0x80, 0x76, 0x71, 0x68, 0x36,
-	0xeb, 0x3c, 0xe6, 0x81, 0x5c, 0x6e, 0xd9, 0x78, 0xfd, 0x9d, 0xd9, 0x24, 0x01, 0x19, 0xfc, 0x57,
-	0x6f, 0x7e, 0xf9, 0x1d, 0x00, 0x00, 0xff, 0xff, 0x0a, 0xd8, 0xa6, 0x8b, 0x87, 0x02, 0x00, 0x00,
+	// 420 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x53, 0xbd, 0xae, 0xd3, 0x30,
+	0x14, 0x4e, 0xb8, 0xe5, 0x4a, 0xd7, 0x77, 0x8b, 0x10, 0xa4, 0x01, 0xc2, 0x55, 0x06, 0x40, 0x48,
+	0xd8, 0xb4, 0xac, 0x88, 0xa1, 0x48, 0xc0, 0x80, 0x10, 0x44, 0x62, 0xe9, 0x52, 0x39, 0xed, 0xa9,
+	0x6b, 0x35, 0xb1, 0xd3, 0xd8, 0x41, 0xed, 0x5b, 0xf0, 0x14, 0x3c, 0x4b, 0xc7, 0x8e, 0x4c, 0x08,
+	0xb5, 0x2f, 0x82, 0x62, 0xbb, 0x3f, 0x12, 0x29, 0x88, 0xcd, 0x27, 0xe7, 0x7c, 0xe7, 0xfb, 0x89,
+	0x8d, 0xc8, 0x04, 0x54, 0x21, 0x15, 0x29, 0x2b, 0x39, 0xe5, 0x39, 0x28, 0xf2, 0xb5, 0x97, 0x81,
+	0xa6, 0x3d, 0xb2, 0xa8, 0xa1, 0x5a, 0x8d, 0xc6, 0x33, 0xca, 0xc5, 0x28, 0xe7, 0x62, 0xae, 0x70,
+	0x59, 0x49, 0x2d, 0x83, 0x7b, 0x16, 0x80, 0xf7, 0x00, 0xec, 0x00, 0xd1, 0x1d, 0x26, 0x99, 0x34,
+	0x33, 0xa4, 0x39, 0xd9, 0xf1, 0xe8, 0x01, 0x93, 0x92, 0xe5, 0x40, 0x68, 0xc9, 0x09, 0x15, 0x42,
+	0x6a, 0xaa, 0xb9, 0x14, 0x6e, 0x59, 0xd4, 0x75, 0x5d, 0x53, 0x65, 0xf5, 0x94, 0x50, 0xb1, 0x72,
+	0xad, 0x17, 0xe7, 0x84, 0x15, 0x72, 0x02, 0xb9, 0xfa, 0x53, 0x59, 0xd4, 0x1d, 0xcb, 0x06, 0x31,
+	0xb2, 0x1a, 0x6c, 0xe1, 0x5a, 0xcf, 0x6c, 0x45, 0x32, 0xaa, 0xc0, 0x3a, 0x3b, 0xac, 0x2b, 0x29,
+	0xe3, 0xc2, 0x88, 0xb2, 0xb3, 0xc9, 0x14, 0x75, 0x3f, 0x37, 0x13, 0x5f, 0x14, 0x54, 0x6f, 0x1a,
+	0x92, 0x0f, 0x5c, 0xcc, 0x53, 0x58, 0xd4, 0xa0, 0x74, 0x10, 0xa0, 0x4e, 0xad, 0xa0, 0x0a, 0xfd,
+	0x1b, 0xff, 0xe9, 0x55, 0x6a, 0xce, 0xc1, 0x43, 0x84, 0xac, 0x18, 0x41, 0x0b, 0x08, 0x6f, 0x99,
+	0xce, 0x95, 0xf9, 0xf2, 0x91, 0x16, 0x10, 0xdc, 0x45, 0x97, 0x9a, 0x56, 0x0c, 0x74, 0x78, 0x61,
+	0x5a, 0xae, 0x4a, 0x86, 0x28, 0x6a, 0xe3, 0x51, 0xa5, 0x14, 0x0a, 0x82, 0x57, 0xa8, 0xd3, 0x78,
+	0x33, 0x44, 0xd7, 0xfd, 0x04, 0x9f, 0x49, 0x1d, 0x1f, 0x90, 0x83, 0xce, 0xfa, 0xe7, 0x23, 0x2f,
+	0x35, 0xa8, 0x64, 0xd9, 0xb6, 0x5b, 0xfd, 0xcd, 0xc4, 0x5b, 0x84, 0x8e, 0x49, 0x18, 0x13, 0xd7,
+	0xfd, 0xc7, 0xd8, 0x85, 0xd8, 0xc4, 0x86, 0x4d, 0x6c, 0x07, 0xde, 0x4f, 0x94, 0x81, 0xdb, 0x97,
+	0x9e, 0x20, 0x93, 0xef, 0x3e, 0xba, 0xdf, 0x4a, 0xed, 0x7c, 0xbd, 0x46, 0xb7, 0xcd, 0x3f, 0x0b,
+	0xfd, 0x9b, 0x8b, 0xff, 0x32, 0x66, 0x61, 0xc1, 0xbb, 0x16, 0x9d, 0x4f, 0xfe, 0xa9, 0xd3, 0x92,
+	0x9f, 0x0a, 0x1d, 0xbc, 0x5f, 0x6f, 0x63, 0x7f, 0xb3, 0x8d, 0xfd, 0x5f, 0xdb, 0xd8, 0xff, 0xb6,
+	0x8b, 0xbd, 0xcd, 0x2e, 0xf6, 0x7e, 0xec, 0x62, 0x6f, 0x88, 0x19, 0xd7, 0xb3, 0x3a, 0xc3, 0x63,
+	0x59, 0xb8, 0xd7, 0xf1, 0x3c, 0xa7, 0x99, 0xda, 0xbf, 0x94, 0xe5, 0xf1, 0x4a, 0xea, 0x55, 0x09,
+	0x2a, 0xbb, 0x34, 0xf7, 0xe6, 0xe5, 0xef, 0x00, 0x00, 0x00, 0xff, 0xff, 0x63, 0x42, 0xc8, 0xc6,
+	0x4b, 0x03, 0x00, 0x00,
+}
+
+func (m *QueryUserChainLinkRequest) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QueryUserChainLinkRequest) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QueryUserChainLinkRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.Target) > 0 {
+		i -= len(m.Target)
+		copy(dAtA[i:], m.Target)
+		i = encodeVarintQueryChainLinks(dAtA, i, uint64(len(m.Target)))
+		i--
+		dAtA[i] = 0x1a
+	}
+	if len(m.ChainName) > 0 {
+		i -= len(m.ChainName)
+		copy(dAtA[i:], m.ChainName)
+		i = encodeVarintQueryChainLinks(dAtA, i, uint64(len(m.ChainName)))
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.User) > 0 {
+		i -= len(m.User)
+		copy(dAtA[i:], m.User)
+		i = encodeVarintQueryChainLinks(dAtA, i, uint64(len(m.User)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *QueryUserChainLinkResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QueryUserChainLinkResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QueryUserChainLinkResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	{
+		size, err := m.Link.MarshalToSizedBuffer(dAtA[:i])
+		if err != nil {
+			return 0, err
+		}
+		i -= size
+		i = encodeVarintQueryChainLinks(dAtA, i, uint64(size))
+	}
+	i--
+	dAtA[i] = 0xa
+	return len(dAtA) - i, nil
 }
 
 func (m *QueryUserChainLinksRequest) Marshal() (dAtA []byte, err error) {
@@ -275,6 +470,38 @@ func encodeVarintQueryChainLinks(dAtA []byte, offset int, v uint64) int {
 	dAtA[offset] = uint8(v)
 	return base
 }
+func (m *QueryUserChainLinkRequest) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.User)
+	if l > 0 {
+		n += 1 + l + sovQueryChainLinks(uint64(l))
+	}
+	l = len(m.ChainName)
+	if l > 0 {
+		n += 1 + l + sovQueryChainLinks(uint64(l))
+	}
+	l = len(m.Target)
+	if l > 0 {
+		n += 1 + l + sovQueryChainLinks(uint64(l))
+	}
+	return n
+}
+
+func (m *QueryUserChainLinkResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = m.Link.Size()
+	n += 1 + l + sovQueryChainLinks(uint64(l))
+	return n
+}
+
 func (m *QueryUserChainLinksRequest) Size() (n int) {
 	if m == nil {
 		return 0
@@ -316,6 +543,235 @@ func sovQueryChainLinks(x uint64) (n int) {
 }
 func sozQueryChainLinks(x uint64) (n int) {
 	return sovQueryChainLinks(uint64((x << 1) ^ uint64((int64(x) >> 63))))
+}
+func (m *QueryUserChainLinkRequest) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQueryChainLinks
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QueryUserChainLinkRequest: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QueryUserChainLinkRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field User", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQueryChainLinks
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.User = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ChainName", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQueryChainLinks
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.ChainName = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Target", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQueryChainLinks
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Target = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQueryChainLinks(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *QueryUserChainLinkResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQueryChainLinks
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QueryUserChainLinkResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QueryUserChainLinkResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Link", wireType)
+			}
+			var msglen int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQueryChainLinks
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				msglen |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			if msglen < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			postIndex := iNdEx + msglen
+			if postIndex < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			if err := m.Link.Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+				return err
+			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQueryChainLinks(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQueryChainLinks
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
 }
 func (m *QueryUserChainLinksRequest) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)